# Improve copy efficiency in block decompression