Currently, `UncompressBlock` can sometimes spend a lot of time copying matches
into the destination buffer. If `mLen` is much larger than `offset`,
which will happen when uncompressing constant or mostly-constant data,
then we can get significant speedup by using a more efficient copying
strategy. A new benchmark, `BenchmarkUncompressConstantBlock`, shows ~100x
faster decompression of a 4KB constant block with this change:

```
name                       old time/op  new time/op  delta
UncompressBlock-4           292ns ± 4%   293ns ± 4%     ~     (p=0.425 n=11+11)
UncompressConstantBlock-4  25.5µs ± 5%   0.3µs ± 4%  -98.93%  (p=0.000 n=11+11)
CompressBlock-4            19.4µs ± 5%  19.3µs ± 5%     ~     (p=0.973 n=11+10)
CompressBlockHC-4          38.2µs ± 6%  38.4µs ± 4%     ~     (p=0.760 n=11+11)
CompressEndToEnd-4         18.2µs ± 2%  18.1µs ± 4%     ~     (p=0.898 n=11+11)
```